--- back/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "backend-test",
-  "version": "1.0.0",
-  "description": "a test app using node, express, sqlize, and mysql for backend api consumption",
-  "main": "server.js",
-  "scripts": {
-    "install": "chcp 65001 && echo MySqlPass= >> .env && chcp 1252",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "prod": "node server.js",
-    "serve": "nodemon server.js"
-  },
-  "keywords": [
-    "nodejs",
-    "express",
-    "sqlize",
-    "mysql",
-    "rest",
-    "api"
-  ],
-  "author": "austin.i.benitez",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "mysql2": "^2.2.5",
-    "sequelize": "^6.3.5",
-    "dotenv": "^8.2.0"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.4"
-  }
-}
