+++ example/front/babel.config.js
