+++ example/front/public/favicon.ico
