+++ example/front/src/assets/logo.png
