+++ example/front/src/components/HelloWorld.vue
