+++ example/front/src/components/Navbar.vue
