+++ example/front/src/http-common.js
