+++ example/front/src/main.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import App from './App.vue'
+import App from './views/App.vue'
 import router from './router'
 
 Vue.config.productionTip = false
