+++ example/front/src/views/App.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import Navbar from "./components/Navbar";
+import Navbar from "../components/Navbar";
 
 export default {
   name: "app",
