+++ example/front/vue.config.js
